+++ demuxer_test.go
@@ -90,7 +90,7 @@ func TestDemuxerNextData(t *testing.T) {
 func TestDemuxerRewind(t *testing.T) {
 	r := bytes.NewReader([]byte("content"))
 	dmx := New(context.Background(), r)
-	dmx.packetPool.add(&Packet{Header: &PacketHeader{PID: 1}})
+	dmx.packetPool.Add(&Packet{Header: &PacketHeader{PID: 1}})
 	dmx.dataBuffer = append(dmx.dataBuffer, &Data{})
 	b := make([]byte, 2)
 	_, err := r.Read(b)