+++ CHANGELOG.md
@@ -2,3 +2,5 @@
 
 ## Unreleased
  - Add serialisation methods for PAT and PMT packets
+ - Make ParseData public
+ - Make `PacketPool` and `(p *PacketPool) Add()` public